[FIXME] Fix criticals/crashes@startup: popover etc
ColorPalette should not set the parent to the whole widget, but just
rather call MenuButton.set_popover() and let that method handle this.

MarkerComboBox was not calling the ctors of its virtual bases ObjectBase
and ExtraVfuncs as it must, so the latter's assumption/assertion failed.

AttrDialog has a manually parented Popover but was not present()ing it.